LayoutTests/fast/js/date-toisostring-expected.txt

Tests for Date.toISOString

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Date.toISOString.call({}) threw exception TypeError: 'undefined' is not an object (evaluating 'Date.toISOString.call').
PASS Date.toISOString.call(0) threw exception TypeError: 'undefined' is not an object (evaluating 'Date.toISOString.call').
PASS new Date(-400).toISOString() is '1969-12-31T23:59:59.600Z'
PASS new Date(0).toISOString() is '1970-01-01T00:00:00.000Z'
PASS new Date('1 January 1500 UTC').toISOString() is '1500-01-01T00:00:00.000Z'
PASS new Date('1 January 2000 UTC').toISOString() is '2000-01-01T00:00:00.000Z'
PASS new Date('1 January 4000 UTC').toISOString() is '4000-01-01T00:00:00.000Z'
PASS new Date('1 January 100000 UTC').toISOString() is '+100000-01-01T00:00:00.000Z'
PASS new Date('1 January -1 UTC').toISOString() is '-000001-01-01T00:00:00.000Z'
PASS new Date('10 March 2000 UTC').toISOString() is '2000-03-10T00:00:00.000Z'
PASS throwsRangeError("new Date(NaN).toISOString()") is true
PASS successfullyParsed is true

TEST COMPLETE
